Integration of Amazon S3 with Project Infrastructure (#8)
### Detailed Changes

#### S3 Bucket

At the core of this update, we have added an S3 bucket. This bucket, identified by the resource "aws_s3_bucket" with the ID "libera_s3_bucket", is used to store and retrieve data. We also specified some bucket attributes like bucket ownership controls and public access blocks.

#### IAM Role and Service Account

In AWS, Identity Access Management (IAM) is used to manage access to AWS services and resources securely. An IAM Role is an entity that defines a set of permissions for making AWS service requests.

Here, we added a new IAM role "s3_access" using the Terraform AWS IAM Module. This role is specifically designed to provide service account access for EKS. AWS EKS uses OpenID Connect (OIDC) to allow Kubernetes to assign AWS IAM roles to Kubernetes service accounts. This feature allows applications to assume an IAM Role and receive AWS temporary credentials to access AWS services.

The service account "s3-access" is created in the "main-project" namespace using "kubectl_manifest". This service account is annotated with the Amazon Resource Name (ARN) of the IAM role that we created. The Kubernetes Service Account, in combination with the IAM role, will allow the pods that use this service account to access the S3 bucket.

#### S3 Access Policy

We've also added an IAM policy called "S3AccessPolicy" that allows all actions ("s3:*") on the resources of our S3 bucket. The policy is then attached to our IAM role using "aws_iam_role_policy_attachment", allowing entities assuming this role to perform operations on the S3 bucket.

### Test Deployment
In addition to the infrastructure changes, we've also introduced a new Kubernetes Deployment for testing our configuration. This deployment is defined in the test_s3.yaml file.

This Deployment, named demo-aws-cli, is designed to confirm that our Service Account setup is correctly providing access to the S3 bucket. Here's how it works:

#### Deployment Configuration

The Deployment creates a pod with a single container running the amazon/aws-cli image, which includes the AWS command-line interface.

The container is configured to use the s3-access Service Account, which is associated with the IAM role that has access to the S3 bucket.

The container then executes a command (on startup) that creates a new file, dummy.txt, and attempts to upload this file to our S3 bucket.

#### Testing Bucket Access

This upload operation serves as a test of our setup. If the IAM role and Service Account have been configured correctly, the container should be able to upload the file to the S3 bucket without any issues. If there are problems with the IAM configuration, the upload operation will fail, and we can use this failure to diagnose the issue.

The AWS CLI command that uploads the file (aws s3 cp) also includes the --acl public-read flag. This flag sets the uploaded file's Access Control List (ACL) to public-read, which allows all users (including unauthenticated ones) to read the file. This means that we can verify the upload's success by attempting to download the file from the S3 bucket ourselves with **https://libera-bucket.s3.amazonaws.com/dummy.txt**